[Feat] PreviewDetailCourseViewController 구현 (#53)
