+++ DATEROAD-iOS/DATEROAD-iOS/Application/SceneDelegate.swift
@@ -20,7 +20,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         // 2.
         self.window = UIWindow(windowScene: windowScene)
         // 3.
-        let navigationController = UINavigationController(rootViewController: TabBarController())
+        let navigationController = UINavigationController(rootViewController: PreviewCourseDetailViewController(viewModel: PreviewCourseDetailViewModel()))
         self.window?.rootViewController = navigationController
         // 4.
         self.window?.makeKeyAndVisible()
