+++ DATEROAD-iOS/DATEROAD-iOS/Global/Resources/Assets.xcassets/CourseDetail/img_preview.imageset/Contents.json
@@ -1,0 +1,12 @@
+{
+  "images" : [
+    {
+      "filename" : "ddd.svg",
+      "idiom" : "universal"
+    }
+  ],
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  }
+}
