+++ DATEROAD-iOS/DATEROAD-iOS/Global/UIComponents/TabBarController.swift
@@ -15,7 +15,7 @@ final class TabBarController: UITabBarController {
     // MARK: - UI Properties
     
     //여기 임의로 아무 뷰컨 넣었어용
-    let homeVC = CourseDetailViewController()
+    let homeVC = PreviewCourseDetailViewController(viewModel: PreviewCourseDetailViewModel())
     
     let courseVC = CourseViewController(courseViewModel: CourseViewModel())
