+++ DATEROAD-iOS/DATEROAD-iOS/Presentation/CourseDetail/Cells/MainContentsCell.swift
@@ -23,7 +23,8 @@ final class MainContentsCell: BaseCollectionViewCell {
     
     override func setLayout() {
         mainTextLabel.snp.makeConstraints {
-            $0.top.bottom.equalToSuperview()
+            $0.top.equalToSuperview()
+            $0.bottom.equalToSuperview().offset(-34)
             $0.leading.trailing.equalToSuperview()
         }
     }
@@ -36,7 +37,7 @@ final class MainContentsCell: BaseCollectionViewCell {
 
             $0.font = UIFont.suit(.body_med_13)
             $0.textColor = UIColor(resource: .drBlack)
-            $0.numberOfLines = 0
+            $0.numberOfLines = 3
         }
     }
