--- DATEROAD-iOS/DATEROAD-iOS/Presentation/CourseDetail/Views/CourseDetailViewController.swift
@@ -1,389 +1,0 @@
-//
-//  CourseDetailViewController.swift
-//  DATEROAD-iOS
-//
-//  Created by 김민서 on 7/1/24.
-//
-
-
-import UIKit
-
-import SnapKit
-import Then
-
-final class CourseDetailViewController: BaseViewController {
-    
-    // MARK: - UI Properties
-    
-    private lazy var mainCollectionView = UICollectionView(frame: .zero, collectionViewLayout: self.makeFlowLayout())
-    
-    private let bottomPageControlView = BottomPageControllView()
-
-    private let previewView = PreviewView()
-    
-    // MARK: - Properties
-    
-    private let viewModel: CourseDetailViewModel
-    
-    private var imageData: [ImageModel] = ImageModel.imageContents.map { ImageModel(image: $0) }
-    
-    private var likeSum: Int = ImageModel.likeSum
-    
-    private var mainContentsData: MainContentsModel = MainContentsModel.mainContents
-    
-    private var timelineData: [TimelineModel] = TimelineModel.timelineContents
-    
-    private var coastData: Int = DateInfoModel.coast
-    
-    private var tagData: [DateInfoModel] = DateInfoModel.tagContents
-    
-    private var currentPage: Int = 0
-    
-    init(viewModel: CourseDetailViewModel = CourseDetailViewModel()) {
-        self.viewModel = viewModel
-        super.init(nibName: nil, bundle: nil)
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    // MARK: - Life Cycle
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        
-        setDelegate()
-        registerCell()
-        //setLeftBackButton()
-    }
-    
-    override func setHierarchy() {
-        super.setHierarchy()
-        view.addSubview(mainCollectionView)
-        //내용 열람 전 뷰
-//        view.addSubview(previewView)
-    }
-    
-    override func setLayout() {
-        super.setLayout()
-        
-        mainCollectionView.contentInsetAdjustmentBehavior = .never
-        mainCollectionView.showsVerticalScrollIndicator = false
-        mainCollectionView.showsHorizontalScrollIndicator = false
-        
-        //스크롤 테스트용
-//        mainCollectionView.isScrollEnabled = false
-    
-        mainCollectionView.snp.makeConstraints {
-            $0.edges.equalToSuperview()
-        }
-        
-    }
-    
-    override func setStyle() {
-        super.setStyle()
-        
-        self.view.backgroundColor = UIColor(resource: .drWhite)
-        self.navigationController?.navigationBar.isHidden = true
-        
-        mainCollectionView.autoresizingMask = [.flexibleWidth, .flexibleHeight]
-    }
-    
-    private func setDelegate() {
-        mainCollectionView.delegate = self
-        mainCollectionView.dataSource = self
-    }
-
-}
-
-private extension CourseDetailViewController {
-    
-    func registerCell() {
-        mainCollectionView.do {
-            $0.register(ImageCarouselCell.self, forCellWithReuseIdentifier: ImageCarouselCell.cellIdentifier)
-            $0.register(TitleInfoCell.self, forCellWithReuseIdentifier: TitleInfoCell.cellIdentifier)
-            $0.register(MainContentsCell.self, forCellWithReuseIdentifier: MainContentsCell.cellIdentifier)
-            $0.register(TimelineInfoCell.self, forCellWithReuseIdentifier: TimelineInfoCell.cellIdentifier)
-            $0.register(CoastInfoCell.self, forCellWithReuseIdentifier: CoastInfoCell.cellIdentifier)
-            $0.register(TagInfoCell.self, forCellWithReuseIdentifier: TagInfoCell.cellIdentifier)
-            $0.register(BringCourseCell.self, forCellWithReuseIdentifier: BringCourseCell.cellIdentifier)
-        
-            $0.register(GradientView.self, forSupplementaryViewOfKind: GradientView.elementKinds, withReuseIdentifier: GradientView.identifier)
-            
-            $0.register(InfoHeaderView.self, forSupplementaryViewOfKind: InfoHeaderView.elementKinds, withReuseIdentifier: InfoHeaderView.identifier)
-            
-            $0.register(BottomPageControllView.self, forSupplementaryViewOfKind: BottomPageControllView.elementKinds, withReuseIdentifier: BottomPageControllView.identifier)
-        }
-    }
-    
-    func makeFlowLayout() -> UICollectionViewCompositionalLayout {
-        return UICollectionViewCompositionalLayout { section, _ -> NSCollectionLayoutSection? in
-            let sectionType = self.viewModel.fetchSection(at: section)
-            
-            switch sectionType {
-            case .imageCarousel:
-                return self.makeImageCarouselLayout()
-            case .titleInfo:
-                return self.makeTitleInfoLayout()
-            case .mainContents:
-                return self.makeMainContentsLayout()
-            case .timelineInfo:
-                return self.makeTimelineInfoLayout()
-            case .coastInfo:
-                return self.makeCoastInfoLayout()
-            case .tagInfo:
-                return self.makeTagInfoLayout()
-            case .bringCourse:
-                return self.makeBringCourseLayout()
-            }
-        }
-    }
-    
-    func makeImageCarouselLayout() -> NSCollectionLayoutSection {
-        let itemSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1), heightDimension: .fractionalHeight(1))
-        let item = NSCollectionLayoutItem(layoutSize: itemSize)
-        
-        let groupSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1), heightDimension: .fractionalWidth(1))
-        let group = NSCollectionLayoutGroup.horizontal(layoutSize: groupSize, subitems: [item])
-        
-        let section = NSCollectionLayoutSection(group: group)
-        section.orthogonalScrollingBehavior = .groupPaging
-        
-        let gradient = makeGradientView()
-        let footer = makeBottomPageControllView()
-        
-        section.boundarySupplementaryItems = [gradient, footer]
-
-        return section
-    }
-    
-    func makeTitleInfoLayout() -> NSCollectionLayoutSection {
-        let itemSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1), heightDimension: .estimated(160))
-        let item = NSCollectionLayoutItem(layoutSize: itemSize)
-        item.contentInsets = NSDirectionalEdgeInsets(top: 0, leading: 16, bottom: 0, trailing: 16)
-        
-        let groupSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1), heightDimension: .estimated(160))
-        let group = NSCollectionLayoutGroup.horizontal(layoutSize: groupSize, subitems: [item])
-        
-        let section = NSCollectionLayoutSection(group: group)
-        
-        return section
-    }
-    
-    func makeMainContentsLayout() -> NSCollectionLayoutSection {
-        let itemSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1), heightDimension: .estimated(50))
-        let item = NSCollectionLayoutItem(layoutSize: itemSize)
-        item.contentInsets = NSDirectionalEdgeInsets(top: 0, leading: 16, bottom: 0, trailing: 16)
-        
-        let groupSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1), heightDimension: .estimated(50))
-        let group = NSCollectionLayoutGroup.horizontal(layoutSize: groupSize, subitems: [item])
-        
-        let section = NSCollectionLayoutSection(group: group)
-        
-        return section
-    }
-    
-    func makeTimelineInfoLayout() -> NSCollectionLayoutSection {
-        let itemSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1), heightDimension: .estimated(70))
-        let item = NSCollectionLayoutItem(layoutSize: itemSize)
-        item.contentInsets = NSDirectionalEdgeInsets(top: 0, leading: 16, bottom: 0, trailing: 16)
-        
-        let groupSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1), heightDimension: .estimated(70))
-        let group = NSCollectionLayoutGroup.horizontal(layoutSize: groupSize, subitems: [item])
-        
-        let section = NSCollectionLayoutSection(group: group)
-        section.contentInsets = NSDirectionalEdgeInsets(top: 0, leading: 0, bottom: 15, trailing: 0)
-        
-        let header = makeHeaderView()
-        section.boundarySupplementaryItems = [header]
-        
-        return section
-    }
-    
-    
-    
-    func makeCoastInfoLayout() -> NSCollectionLayoutSection {
-        let itemSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1), heightDimension: .fractionalHeight(1))
-        let item = NSCollectionLayoutItem(layoutSize: itemSize)
-        item.contentInsets = NSDirectionalEdgeInsets(top: 0, leading: 16, bottom: 0, trailing: 16)
-        
-        let groupSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1), heightDimension: .absolute(50))
-        let group = NSCollectionLayoutGroup.horizontal(layoutSize: groupSize, subitems: [item])
-        
-        let section = NSCollectionLayoutSection(group: group)
-        section.contentInsets = NSDirectionalEdgeInsets(top: 0, leading: 0, bottom: 30, trailing: 0)
-        
-        let header = makeHeaderView()
-        section.boundarySupplementaryItems = [header]
-        
-        return section
-    }
-    
-    func makeTagInfoLayout() -> NSCollectionLayoutSection {
-        let itemSize = NSCollectionLayoutSize(widthDimension: .estimated(100), heightDimension: .fractionalHeight(1))
-        let item = NSCollectionLayoutItem(layoutSize: itemSize)
-        
-        let groupSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1), heightDimension: .absolute(30))
-        let group = NSCollectionLayoutGroup.horizontal(layoutSize: groupSize, subitems: [item])
-        group.contentInsets = NSDirectionalEdgeInsets(top: 0, leading: 16, bottom: 0, trailing: 0)
-        group.interItemSpacing = .fixed(7)
-        
-        let section = NSCollectionLayoutSection(group: group)
-        section.contentInsets = NSDirectionalEdgeInsets(top: 0, leading: 0, bottom: 61, trailing: 0)
-        
-        let header = makeHeaderView()
-        section.boundarySupplementaryItems = [header]
-        
-        return section
-    }
-    
-    
-    func makeBringCourseLayout() -> NSCollectionLayoutSection {
-        let itemSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1), heightDimension: .fractionalHeight(1))
-        let item = NSCollectionLayoutItem(layoutSize: itemSize)
-        item.contentInsets = NSDirectionalEdgeInsets(top: 0, leading: 16, bottom: 0, trailing: 16)
-        
-        let groupSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1), heightDimension: .absolute(54))
-        let group = NSCollectionLayoutGroup.horizontal(layoutSize: groupSize, subitems: [item])
-        
-        let section = NSCollectionLayoutSection(group: group)
-        section.contentInsets = NSDirectionalEdgeInsets(top: 0, leading: 0, bottom: 30, trailing: 0)
-        
-        return section
-    }
-    
-    func makeGradientView() -> NSCollectionLayoutBoundarySupplementaryItem {
-        let gradientSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1), heightDimension: .fractionalWidth(0.27))
-        let collectionViewWidth = mainCollectionView.frame.width
-        let gradient = NSCollectionLayoutBoundarySupplementaryItem(layoutSize: gradientSize, elementKind: GradientView.elementKinds, alignment: .top, absoluteOffset: CGPoint(x: 0, y: collectionViewWidth * 0.27))
-      
-        return gradient
-    }
-    
-    func makeHeaderView() -> NSCollectionLayoutBoundarySupplementaryItem {
-        let headerSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1), heightDimension: .absolute(37))
-        let header = NSCollectionLayoutBoundarySupplementaryItem(layoutSize: headerSize, elementKind: InfoHeaderView.elementKinds, alignment: .top)
-        header.contentInsets = NSDirectionalEdgeInsets(top: 0, leading: 16, bottom: 12, trailing: 16)
-        return header
-    }
-    
-    func makeBottomPageControllView() -> NSCollectionLayoutBoundarySupplementaryItem {
-        let footerSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1), heightDimension: .estimated(22))
-        let footer = NSCollectionLayoutBoundarySupplementaryItem(layoutSize: footerSize, elementKind: BottomPageControllView.elementKinds, alignment: .bottom, absoluteOffset: CGPoint(x: 0, y: -33))
-        footer.contentInsets = NSDirectionalEdgeInsets(top: 0, leading: 16, bottom: 0, trailing: 16)
-        return footer
-    }
-}
-
-extension CourseDetailViewController: ImageCarouselDelegate {
-    
-    func didSwipeImage(index: Int, vc: UIPageViewController, vcData: [UIViewController]) {
-        currentPage = index
-        if let bottomPageControllView = mainCollectionView.supplementaryView(forElementKind: BottomPageControllView.elementKinds, at: IndexPath(item: 0, section: 0)) as? BottomPageControllView {
-            bottomPageControllView.pageIndex = currentPage
-        }
-    }
-}
-
-extension CourseDetailViewController: UICollectionViewDelegate, UICollectionViewDataSource {
-    func numberOfSections(in collectionView: UICollectionView) -> Int {
-        return viewModel.numberOfSections
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        let sectionType = viewModel.fetchSection(at: section)
-        
-        switch sectionType {
-        case .imageCarousel:
-            return viewModel.imageCarouselViewModel.numberOfItems
-        case .timelineInfo:
-            return timelineData.count
-        case .tagInfo:
-            return tagData.count
-        default:
-            return viewModel.numberOfItemsInSection(section)
-        }
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        
-        let sectionType = viewModel.fetchSection(at: indexPath.section)
-        
-        switch sectionType {
-        case .imageCarousel:
-            guard let imageCarouselCell = collectionView.dequeueReusableCell(withReuseIdentifier: ImageCarouselCell.cellIdentifier, for: indexPath) as? ImageCarouselCell else {
-                fatalError("Unable to dequeue ImageCarouselCell")
-            }
-            imageCarouselCell.setPageVC(imageData: imageData)
-            imageCarouselCell.delegate = self
-            return imageCarouselCell
-        case .titleInfo:
-            guard let titleInfoCell = collectionView.dequeueReusableCell(withReuseIdentifier: TitleInfoCell.cellIdentifier, for: indexPath) as? TitleInfoCell else {
-                fatalError("Unable to dequeue MainContentsCell")
-            }
-            titleInfoCell.backgroundColor = .red
-            titleInfoCell.setCell(mainContentsData: mainContentsData)
-            return titleInfoCell
-        case .mainContents:
-            guard let mainContentsCell = collectionView.dequeueReusableCell(withReuseIdentifier: MainContentsCell.cellIdentifier, for: indexPath) as? MainContentsCell else {
-                fatalError("Unable to dequeue MainContentsCell")
-            }
-            mainContentsCell.backgroundColor = .blue
-            mainContentsCell.setCell(mainContentsData: mainContentsData)
-            return mainContentsCell
-        case .timelineInfo:
-            guard let timelineInfoCell = collectionView.dequeueReusableCell(withReuseIdentifier: TimelineInfoCell.cellIdentifier, for: indexPath) as? TimelineInfoCell else {
-                fatalError("Unable to dequeue TimelineInfoCell")
-            }
-            timelineInfoCell.setCell(timelineData[indexPath.row])
-            return timelineInfoCell
-        case .coastInfo:
-            guard let coastInfoCell = collectionView.dequeueReusableCell(withReuseIdentifier: CoastInfoCell.cellIdentifier, for: indexPath) as? CoastInfoCell else {
-                fatalError("Unable to dequeue CoastInfoCell")
-            }
-            coastInfoCell.setCell(coastData: coastData)
-            return coastInfoCell
-        case .tagInfo:
-            guard let tagInfoCell = collectionView.dequeueReusableCell(withReuseIdentifier: TagInfoCell.cellIdentifier, for: indexPath) as? TagInfoCell else {
-                fatalError("Unable to dequeue TagInfoCell")
-            }
-            tagInfoCell.setCell(tagData: tagData[indexPath.row])
-            return tagInfoCell
-        case .bringCourse:
-            guard let bringCourseCell = collectionView.dequeueReusableCell(withReuseIdentifier: BringCourseCell.cellIdentifier, for: indexPath) as? BringCourseCell else {
-                fatalError("Unable to dequeue TagInfoCell")
-            }
-            return bringCourseCell
-        }
-        
-        
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, viewForSupplementaryElementOfKind kind: String, at indexPath: IndexPath) -> UICollectionReusableView {
-        if kind == InfoHeaderView.elementKinds {
-            guard let header = collectionView.dequeueReusableSupplementaryView(ofKind: kind, withReuseIdentifier: InfoHeaderView.identifier, for: indexPath) as? InfoHeaderView else { return UICollectionReusableView() }
-            switch viewModel.fetchSection(at: indexPath.section) {
-            case .timelineInfo:
-                header.bindTitle(headerTitle: StringLiterals.CourseDetail.timelineInfoLabel)
-            case .coastInfo:
-                header.bindTitle(headerTitle: StringLiterals.CourseDetail.coastInfoLabel)
-            case .tagInfo:
-                header.bindTitle(headerTitle: StringLiterals.CourseDetail.tagInfoLabel)
-            default:
-                break
-            }
-            return header
-        } else if kind == BottomPageControllView.elementKinds {
-            guard let footer = collectionView.dequeueReusableSupplementaryView(ofKind: kind, withReuseIdentifier: BottomPageControllView.identifier, for: indexPath) as? BottomPageControllView else { return UICollectionReusableView() }
-            //이미지 갯수가 총 인덱스
-            footer.pageIndexSum = imageData.count
-            return footer
-        } else if kind == GradientView.elementKinds {
-            guard let gradient = collectionView.dequeueReusableSupplementaryView(ofKind: kind, withReuseIdentifier: GradientView.identifier, for: indexPath) as? GradientView else { return UICollectionReusableView() }
-            return gradient
-        } else {
-            return UICollectionReusableView()
-        }
-    }
-}
