+++ DATEROAD-iOS/DATEROAD-iOS/Presentation/MyPage/Views/MyPageViewController.swift
@@ -156,7 +156,7 @@ extension MyPageViewController: UITableViewDelegate {
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
         switch MyPageSection.dataSource[indexPath.item] {
         case .myCourse:
-            let myCourseVC = CourseDetailViewController(viewModel: CourseDetailViewModel())
+            let myCourseVC = PreviewCourseDetailViewController(viewModel: PreviewCourseDetailViewModel())
             self.navigationController?.pushViewController(myCourseVC, animated: false)
         case .pointSystem:
             let pointSystemVC = PointSystemViewController(pointSystemViewModel: PointSystemViewModel())
