+++ DATEROAD-iOS/DATEROAD-iOS/Presentation/PreviewCourseDetail/Views/ContentMaskView.swift
@@ -1,8 +1,8 @@
 //
-//  PreviewViewController.swift
+//  ContentMaskView.swift
 //  DATEROAD-iOS
 //
-//  Created by 김민서 on 7/7/24.
+//  Created by 김민서 on 7/12/24.
 //
 
 import UIKit
@@ -10,7 +10,7 @@ import UIKit
 import SnapKit
 import Then
 
-class PreviewView: BaseView {
+class ContentMaskView: UICollectionReusableView{
     
     // MARK: - UI Properties
     
@@ -18,63 +18,77 @@ class PreviewView: BaseView {
     
     private let mainTitleLabel = UILabel()
     
+    private let pointImageView = UIImageView(image: .imgPreview)
+    
     private let subTitleLabel = UILabel()
     
     private let readCourseButton = UIButton()
     
     private let gradient = CAGradientLayer()
     
+    // MARK: - Properties
+    
+    static let elementKinds: String = "ContentMaskView"
+    
+    static let identifier: String = "ContentMaskView"
     // MARK: - Life Cycle
     
     override init(frame: CGRect) {
         super.init(frame: frame)
+        
+        setHierarchy()
+        setLayout()
+        setStyle()
     }
     
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
     
-    override func setHierarchy() {
-        super.setHierarchy()
+    func setHierarchy() {
         
         self.addSubviews(
             gradientView,
+            pointImageView,
             mainTitleLabel,
             subTitleLabel,
             readCourseButton
         )
     }
     
-    override func setLayout() {
-        super.setLayout()
-        
+    func setLayout() {
+
+        self.backgroundColor = .white
         gradientView.snp.makeConstraints {
-            $0.bottom.horizontalEdges.equalToSuperview()
-            $0.height.equalTo(233)
+            $0.edges.equalToSuperview()
+        }
+        
+        pointImageView.snp.makeConstraints {
+            $0.top.equalToSuperview().inset(72)
+            $0.centerX.equalToSuperview()
+            $0.width.equalTo(134)
+            $0.height.equalTo(100)
         }
         
         mainTitleLabel.snp.makeConstraints {
-            $0.bottom.equalTo(subTitleLabel.snp.top).offset(-4)
+            $0.top.equalTo(pointImageView.snp.bottom).offset(16)
             $0.centerX.equalTo(gradientView)
         }
         
         subTitleLabel.snp.makeConstraints {
-            $0.bottom.equalTo(readCourseButton.snp.top).offset(-25)
+            $0.top.equalTo(mainTitleLabel.snp.bottom).offset(4)
             $0.centerX.equalToSuperview()
         }
         
         readCourseButton.snp.makeConstraints {
-            $0.bottom.equalToSuperview().inset(38)
+            $0.top.equalTo(subTitleLabel.snp.bottom).offset(25)
             $0.horizontalEdges.equalToSuperview().inset(60)
             $0.height.equalTo(54)
         }
     }
     
-    override func setStyle() {
-        super.setStyle()
-        
-        gradientView.backgroundColor = .red
-        
+    func setStyle() {
+
         mainTitleLabel.do {
             $0.text = "코스 정보가 궁금하신가요?"
             $0.textColor = UIColor(resource: .drBlack)
